building initial input file

+++ PageRankMapper.java
@@ -5,7 +5,7 @@ import org.apache.hadoop.conf.Configuration;
 import org.apache.hadoop.conf.Configured;
 import org.apache.hadoop.fs.Path;
 import org.apache.hadoop.fs.FileSystem;
-import org.apache.hadoop.io.LongWritable;
+import org.apache.hadoop.io.FloatWritable;
 import org.apache.hadoop.io.Text;
 import org.apache.hadoop.mapreduce.Job;
 import org.apache.hadoop.mapreduce.lib.input.FileInputFormat;
@@ -16,12 +16,12 @@ import org.apache.hadoop.util.Tool;
 import org.apache.hadoop.util.ToolRunner;
 import org.apache.hadoop.mapreduce.Mapper;
 
-public class PageRankMapper extends Mapper<LongWritable, Text, Text, LongWritable> {
+public class PageRankMapper extends Mapper<FloatWritable, Text, Text, FloatWritable> {
 
   private Text word = new Text();
 
   @Override
-  public void map(LongWritable key, Text value,
+  public void map(FloatWritable key, Text value,
                   Context context) throws IOException, InterruptedException {
     String line = value.toString();
     List<String> fields = Arrays.asList(line.split("\\s+"));
@@ -32,8 +32,8 @@ public class PageRankMapper extends Mapper<LongWritable, Text, Text, LongWritabl
     for (String outlinkId : fields.subList(2, fields.size() - 1))
     {
       word.set(outlinkId);
-      //output.collect(word, new LongWritable(portion));
-      context.write(word, new LongWritable(portion));
+      //output.collect(word, new FloatWritable(portion));
+      context.write(word, new FloatWritable(portion));
     }
     // parse nid, init_pagerank, and outlinks from line
     // for nid in outlinks:
